refactor, create getToken and getMetadata to reuse it in both methods
