--- vclogin/pages/api/new.ts
@@ -1,16 +1,0 @@
-import { NextApiRequest, NextApiResponse } from "next";
-import { Redis } from "ioredis";
-
-var redis: Redis;
-try {
-  redis = new Redis(parseInt(process.env.REDIS_PORT!), process.env.REDIS_HOST!);
-} catch (error) {
-  console.error("Failed to connect to Redis:", error);
-}
-
-export default async function handler(
-  req: NextApiRequest,
-  res: NextApiResponse<any>,
-) {
-  return res.status(200).json({ message: "Hello World" });
-}
